refactor(cli): move cli() adapter to @routecraft/tools, fix review bugs
- Create @routecraft/tools package with cli() source/destination adapters
  moved out of @routecraft/routecraft core (system-native tooling belongs
  in its own package, paralleling @routecraft/ai for AI-native tooling)
- Fix double "craft run" prefix in per-command help output
- Fix double error output in cliRunner for unknown commands
- Fix craft('greet') in docs (craft() takes no args, use craft().id())
- Deduplicate collectDefinitions logic in cliRunner
- Update all imports in @routecraft/cli, tests, and docs to use
  @routecraft/tools
