+++ apps/routecraft.dev/src/app/docs/advanced/build-a-cli/page.md
@@ -13,7 +13,7 @@ For standalone packaging, the `cliRunner()` helper lets you bypass `craft run` e
 ## Install
 
 ```bash
-npm install @routecraft/routecraft @routecraft/cli zod
+npm install @routecraft/routecraft @routecraft/tools @routecraft/cli zod
 ```
 
 Any [Standard Schema](https://standardschema.dev)-compliant library works in place of Zod (Valibot, ArkType, etc.).
@@ -24,7 +24,8 @@ Each `cli()` call registers one command. Pass a schema to define flags and a des
 
 ```ts
 // mycli.ts
-import { craft, cli } from '@routecraft/routecraft';
+import { craft } from '@routecraft/routecraft';
+import { cli } from '@routecraft/tools';
 import { z } from 'zod';
 
 export default [
@@ -72,7 +73,8 @@ Use `cliRunner()` to run routes directly without `craft run`. Import it from `@r
 ```ts
 #!/usr/bin/env tsx
 // mycli.ts
-import { craft, cli } from '@routecraft/routecraft';
+import { craft } from '@routecraft/routecraft';
+import { cli } from '@routecraft/tools';
 import { cliRunner } from '@routecraft/cli/runner';
 import { z } from 'zod';
