+++ packages/cli/package.json
@@ -28,6 +28,7 @@
   "dependencies": {
     "@opentelemetry/sdk-trace-base": "^2.6.0",
     "@routecraft/routecraft": "^0.5.0",
+    "@routecraft/tools": "workspace:^0.5.0",
     "agent-browser": "^0.17.1",
     "better-sqlite3": "^11.10.0",
     "cheerio": "^1.2.0",
