+++ packages/tools/src/cli/source.ts
@@ -1,7 +1,7 @@
-import type { Exchange, ExchangeHeaders } from "../../exchange";
-import type { Source } from "../../operations/from";
-import type { CraftContext } from "../../context";
-import { rcError } from "../../error";
+import type { Exchange, ExchangeHeaders } from "@routecraft/routecraft";
+import type { Source } from "@routecraft/routecraft";
+import type { CraftContext } from "@routecraft/routecraft";
+import { rcError } from "@routecraft/routecraft";
 import type { CliServerOptions } from "./types";
 import {
   ADAPTER_CLI_ARGS,
