+++ packages/tools/test/cli.test.ts
@@ -1,16 +1,15 @@
 import { describe, test, expect, afterEach, vi } from "vitest";
 import { z } from "zod";
 import { testContext, type TestContext } from "@routecraft/testing";
+import { craft, noop } from "@routecraft/routecraft";
 import {
-  craft,
   cli,
-  noop,
   ADAPTER_CLI_ARGS,
   ADAPTER_CLI_REGISTRY,
   isCliSource,
   parseFlags,
   extractJsonSchema,
-} from "@routecraft/routecraft";
+} from "@routecraft/tools";
 
 /**
  * CLI adapter tests
